refactor node wrapper class
